requirements.txt
fastapi
uvicorn
pydantic
sentence-transformers
chromadb
openai
numpy
python-dotenv